pyproject.toml
[project]
name = "eth-protocol-expert"
version = "0.1.0"
description = "Ethereum Protocol Intelligence System"
requires-python = ">=3.11,<3.14"

dependencies = [
    # Database
    "asyncpg>=0.29.0",
    "pgvector>=0.2.4",

    # Embeddings
    "voyageai>=0.3.0rc0",

    # LLM
    "anthropic>=0.40.0",

    # NLI validation (Phase 2)
    "transformers>=4.36.0",
    "torch>=2.1.0",

    # Reranking (Phase 3)
    "cohere>=5.0.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",

    # Graph database (Phase 4)
    "falkordb>=1.0.0",

    # Utilities
    "httpx>=0.26.0",
    "pyyaml>=6.0.0",
    "tiktoken>=0.5.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
